5547 - Move admin links redux to store/links
