--- src/client/store/admin/slice/name.ts
@@ -1,1 +1,0 @@
-export const AdminSliceName = 'admin'
